sghs

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int LLI;
 	float F;
 
-	printf("Size of a char: %lu byte(s)\n"(unsigned long)sizeof(C));
-	printf("Size of an int: %lu byte(s)\n"(unsigned long)sizeof(I));
-	ptintf("Size of a long int: %lu byte(s)\n"(unsigned long)sizeof(LI));
-	printf("Size of a long long int: %lu byte(s)\n"(unsigned long)sizeof(LLI));
-	printf("Size of a float: %lu byte(s)\n"(unsigned long)sizeof(F));
+	printf("Size of a char: %lu byte(s)\n"sizeof(C));
+	printf("Size of an int: %lu byte(s)\n"sizeof(I));
+	ptintf("Size of a long int: %lu byte(s)\n"sizeof(LI));
+	printf("Size of a long long int: %lu byte(s)\n"sizeof(LLI));
+	printf("Size of a float: %lu byte(s)\n"sizeof(F));
 	return (0);
 }
